moving r4, r5 specific directories to a new direcotory named versions to maintain the consistency
