+++ components/fhir-core/src/main/java/org/wso2/healthcare/codegen/tool/framework/fhir/core/versions/r4/model/FHIRR4TerminologyDef.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.healthcare.codegen.tool.framework.fhir.core.r4.model;
+package org.wso2.healthcare.codegen.tool.framework.fhir.core.versions.r4.model;
 
 import org.hl7.fhir.r4.model.Resource;
 import org.wso2.healthcare.codegen.tool.framework.commons.model.SpecModel;
