+++ components/fhir-core/src/main/java/org/wso2/healthcare/codegen/tool/framework/fhir/core/versions/r4/oas/R4OASGenUtils.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.healthcare.codegen.tool.framework.fhir.core.r4.oas;
+package org.wso2.healthcare.codegen.tool.framework.fhir.core.versions.r4.oas;
 
 import io.swagger.v3.core.util.Json;
 import io.swagger.v3.core.util.Yaml;
