+++ components/fhir-core/src/main/java/org/wso2/healthcare/codegen/tool/framework/fhir/core/versions/r4/util/R4CardinalityTypes.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.healthcare.codegen.tool.framework.fhir.core.r4.util;
+package org.wso2.healthcare.codegen.tool.framework.fhir.core.versions.r4.util;
 
 import java.util.regex.Pattern;
